bug fix and refactor to remove top
auth had to be imported to app to get the login functions. instead, these were moved into
app.py and top.py is removed as its no longer necessary.
